Bug: 792055 Reviewer: none Comment: In the Wizard - Do not allow Ports below 1024

+++ trunk/ssf/ssf-install-config/src/org/kabling/teaming/install/client/widgets/SpinnerListener.java
@@ -29,5 +29,5 @@ public interface SpinnerListener
 	 * @param value
 	 *            the current spinner value
 	 */
-	void onSpinning(long value);
+	void onSpinning(double value);
 }
